requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
numpy>=1.26.0
huggingface_hub>=0.20.0
python-dotenv>=1.0.0